fixed the fusion rrf build and exclude the wasm build (#126)

+++ src/functions/scalar/fusion_rrf/implementation.cpp
@@ -6,7 +6,7 @@ namespace flockmtl {
 // Different entries with the same RRF score are assigned different, consecutive, rankings arbitrarily
 std::vector<double> FusionRRF::Operation(duckdb::DataChunk& args) {
     // recommended rrf constant is 60
-    long rrf_constant = 60;
+    int32_t rrf_constant = 60;
     int num_different_scores = static_cast<int>(args.ColumnCount());
     int num_entries = static_cast<int>(args.size());
 
@@ -16,17 +16,17 @@ std::vector<double> FusionRRF::Operation(duckdb::DataChunk& args) {
     // for each column (scoring system), we want a vector of individual input scores
     for (int i = 0; i < num_different_scores; i++) {
         // extract a single column's rankings. There should be no null values
-        std::vector<long> extracted_ranks(num_entries);
+        std::vector<int32_t> extracted_ranks(num_entries);
         for (int j = 0; j < num_entries; j++) {
             auto valueWrapper = args.data[i].GetValue(j);
             if (!valueWrapper.IsNull()) {
-                long value = valueWrapper.GetValue<long>();
+                int32_t value = valueWrapper.GetValue<int32_t>();
                 extracted_ranks[j] = value;
             }
         }
 
-        // If all entries have the same score or are NULL (0), then this scoring system can be considered useless and should be ignored
-        // Or else, all entries would get assigned the best rank possible, even if they are 0
+        // If all entries have the same score or are NULL (0), then this scoring system can be considered useless and
+        // should be ignored Or else, all entries would get assigned the best rank possible, even if they are 0
         if (std::all_of(extracted_ranks.begin(), extracted_ranks.end(), [](const auto& rank) { return rank <= 1; })) {
             // if there is only one entry and the rank isn't 0, it's a valid ranking. We don't want to skip.
             if (!(num_entries == 1 && extracted_ranks[0] == 1)) {
