+++ test/functions/scalar/fusion_rrf/unittest_fusion_rrf.cpp
@@ -9,7 +9,7 @@ TEST_CASE("Unit test for flockmtl::FusionRRF with 2 DOUBLES", "[fusion_rrf][floc
 
     // Create a DataChunk and initialize it with the default allocator
     duckdb::DataChunk chunk;
-    auto &allocator = duckdb::Allocator::DefaultAllocator();
+    auto& allocator = duckdb::Allocator::DefaultAllocator();
     // Initialize with capacity 1 (one row)
     chunk.Initialize(allocator, types, 1);
 
@@ -25,7 +25,7 @@ TEST_CASE("Unit test for flockmtl::FusionRRF with 2 DOUBLES", "[fusion_rrf][floc
 
     // Verify the result
     REQUIRE(result.size() == 1);
-    REQUIRE(result[0] == 1.0/61 + 1.0/61);
+    REQUIRE(result[0] == 1.0 / 61 + 1.0 / 61);
 }
 
 TEST_CASE("Unit test for flockmtl::FusionRRF with 2 columns", "[fusion_rrf][flockmtl]") {
@@ -34,7 +34,7 @@ TEST_CASE("Unit test for flockmtl::FusionRRF with 2 columns", "[fusion_rrf][floc
 
     // Create a DataChunk and initialize it with the default allocator
     duckdb::DataChunk chunk;
-    auto &allocator = duckdb::Allocator::DefaultAllocator();
+    auto& allocator = duckdb::Allocator::DefaultAllocator();
     // Initialize with capacity 5 (five rows)
     chunk.Initialize(allocator, types, 5);
 
@@ -42,8 +42,8 @@ TEST_CASE("Unit test for flockmtl::FusionRRF with 2 columns", "[fusion_rrf][floc
     chunk.SetCardinality(5);
 
     // Data we will use to populate the DataChunk
-    constexpr std::array<long, 5> bm25_scores = {1, 2, 3, 4, 5};
-    constexpr std::array<long, 5> vs_scores = {4, 2, 1, 5, 3};
+    constexpr std::array<int32_t, 5> bm25_scores = {1, 2, 3, 4, 5};
+    constexpr std::array<int32_t, 5> vs_scores = {4, 2, 1, 5, 3};
 
     // Populate the DataChunk with test data
     for (size_t i = 0; i < bm25_scores.size(); ++i) {
@@ -55,7 +55,8 @@ TEST_CASE("Unit test for flockmtl::FusionRRF with 2 columns", "[fusion_rrf][floc
     const std::vector<double> result = flockmtl::FusionRRF::Operation(chunk);
 
     // Verify the result
-    constexpr std::array<double, 5> expected_results = {(1.0/61 + 1.0/64), (2.0/62), (1.0/63 + 1.0/61), (1.0/64 + 1.0/65), (1.0/65 + 1.0/63)};
+    constexpr std::array<double, 5> expected_results = {(1.0 / 61 + 1.0 / 64), (2.0 / 62), (1.0 / 63 + 1.0 / 61),
+                                                        (1.0 / 64 + 1.0 / 65), (1.0 / 65 + 1.0 / 63)};
     REQUIRE(result.size() == expected_results.size());
     for (size_t i = 0; i < expected_results.size(); ++i) {
         REQUIRE(result[i] == expected_results[i]);
@@ -64,11 +65,12 @@ TEST_CASE("Unit test for flockmtl::FusionRRF with 2 columns", "[fusion_rrf][floc
 
 TEST_CASE("Unit test for flockmtl::FusionRRF with 3 columns", "[fusion_rrf][flockmtl]") {
     // Define the column types (2 DOUBLE columns)
-    const duckdb::vector<duckdb::LogicalType> types = {duckdb::LogicalType::BIGINT, duckdb::LogicalType::BIGINT, duckdb::LogicalType::BIGINT};
+    const duckdb::vector<duckdb::LogicalType> types = {duckdb::LogicalType::BIGINT, duckdb::LogicalType::BIGINT,
+                                                       duckdb::LogicalType::BIGINT};
 
     // Create a DataChunk and initialize it with the default allocator
     duckdb::DataChunk chunk;
-    auto &allocator = duckdb::Allocator::DefaultAllocator();
+    auto& allocator = duckdb::Allocator::DefaultAllocator();
     // Initialize with capacity 5 (five rows)
     chunk.Initialize(allocator, types, 5);
 
@@ -76,9 +78,9 @@ TEST_CASE("Unit test for flockmtl::FusionRRF with 3 columns", "[fusion_rrf][floc
     chunk.SetCardinality(5);
 
     // Data we will use to populate the DataChunk
-    constexpr std::array<long, 5> bm25_scores = {1, 2, 3, 4, 5};
-    constexpr std::array<long, 5> vs_scores = {4, 2, 1, 5, 3};
-    constexpr std::array<long, 5> random_scores = {5, 4, 3, 2, 1};
+    constexpr std::array<int32_t, 5> bm25_scores = {1, 2, 3, 4, 5};
+    constexpr std::array<int32_t, 5> vs_scores = {4, 2, 1, 5, 3};
+    constexpr std::array<int32_t, 5> random_scores = {5, 4, 3, 2, 1};
 
     // Populate the DataChunk with test data
     for (size_t i = 0; i < bm25_scores.size(); ++i) {
@@ -93,7 +95,8 @@ TEST_CASE("Unit test for flockmtl::FusionRRF with 3 columns", "[fusion_rrf][floc
     // Verify the result
     std::vector<double> expected_results(5);
     for (int i = 0; i < 5; ++i) {
-        expected_results[i] = 1.0 / (60.0 + bm25_scores[i]) + 1.0 / (60.0 + vs_scores[i]) + 1.0 / (60.0 + random_scores[i]);
+        expected_results[i] =
+            1.0 / (60.0 + bm25_scores[i]) + 1.0 / (60.0 + vs_scores[i]) + 1.0 / (60.0 + random_scores[i]);
     }
     REQUIRE(result.size() == expected_results.size());
     for (size_t i = 0; i < expected_results.size(); ++i) {
@@ -101,14 +104,15 @@ TEST_CASE("Unit test for flockmtl::FusionRRF with 3 columns", "[fusion_rrf][floc
     }
 }
 
-// Entire columns of first place ranking essentially mean the scoring source did not rank anything and should be ignored.
+// Entire columns of first place ranking essentially mean the scoring source did not rank anything and should be
+// ignored.
 TEST_CASE("Unit test for flockmtl::FusionRRF with entire column of first place ranking", "[fusion_rrf][flockmtl]") {
     // Define the column types (2 DOUBLE columns)
     const duckdb::vector<duckdb::LogicalType> types = {duckdb::LogicalType::BIGINT, duckdb::LogicalType::BIGINT};
 
     // Create a DataChunk and initialize it with the default allocator
     duckdb::DataChunk chunk;
-    auto &allocator = duckdb::Allocator::DefaultAllocator();
+    auto& allocator = duckdb::Allocator::DefaultAllocator();
     // Initialize with capacity 5 (five rows)
     chunk.Initialize(allocator, types, 5);
 
@@ -116,7 +120,7 @@ TEST_CASE("Unit test for flockmtl::FusionRRF with entire column of first place r
     chunk.SetCardinality(5);
 
     // Data we will use to populate the DataChunk
-    constexpr std::array<long, 5> vs_scores = {3, 1, 5, 2, 4};
+    constexpr std::array<int32_t, 5> vs_scores = {3, 1, 5, 2, 4};
 
     // Populate the DataChunk with test data
     for (size_t i = 0; i < vs_scores.size(); ++i) {
@@ -128,7 +132,7 @@ TEST_CASE("Unit test for flockmtl::FusionRRF with entire column of first place r
     const std::vector<double> result = flockmtl::FusionRRF::Operation(chunk);
 
     // Verify the result
-    constexpr std::array<double, 5> expected_results = {1.0/63, 1.0/61, 1.0/65, 1.0/62, 1.0/64};
+    constexpr std::array<double, 5> expected_results = {1.0 / 63, 1.0 / 61, 1.0 / 65, 1.0 / 62, 1.0 / 64};
     REQUIRE(result.size() == expected_results.size());
     for (size_t i = 0; i < expected_results.size(); ++i) {
         REQUIRE(result[i] == expected_results[i]);
